AI_microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0